+++ ocaml-lsp-server/src/code_actions/action_destruct.mli
@@ -1,1 +1,5 @@
-val t : Code_action.t
+open Import
+
+val kind : CodeActionKind.t
+
+val t : State.t -> Code_action.t